requirements.txt
psycopg2-binary
pandas
boto3
selenium
python-dotenv

